discharge variable submodel
