+++ docs/source/api/models/submodels/external_circuit/discharge_throughput.rst
@@ -1,0 +1,7 @@
+Discharge and throughput variables
+==================================
+
+Calculates the discharge and throughput variables (capacity and power) for the battery.
+
+.. autoclass:: pybamm.external_circuit.DischargeThroughput
+    :members:
