+++ pybamm/models/full_battery_models/base_battery_model.py
@@ -1160,6 +1160,9 @@ class BaseBatteryModel(pybamm.BaseModel):
                 self.param, self.options["operating mode"], self.options
             )
         self.submodels["external circuit"] = model
+        self.submodels["discharge and throughput variables"] = (
+            pybamm.external_circuit.DischargeThroughput(self.param, self.options)
+        )
 
     def set_transport_efficiency_submodels(self):
         self.submodels["electrolyte transport efficiency"] = (
